refactor: resolve circular import via import in function
